% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{defaultInfoPanel}
\alias{defaultInfoPanel}
\alias{timeSeriesPanel}
\title{Info panels for the built-in viewer}
\usage{
defaultInfoPanel(point, data, labels)

timeSeriesPanel(point, data, labels)
}
\arguments{
\item{point}{A numeric vector giving the current point location.}

\item{data}{A list of data values for each image at the current point.
Note that, for images of more than three dimensions, there will be more
than one value per image.}

\item{labels}{A character vector of image labels.}
}
\description{
A default info panel for \code{\link{view}}, which shows the labels and
values of each image at the current point, and a panel suitable for
plotting four-dimensional time series images.
}
\seealso{
\code{\link{view}}
}
\author{
Jon Clayden <code@clayden.org>
}
