% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{data}
\name{ExtensionCodes}
\alias{ExtensionCodes}
\title{NIfTI extension codes}
\format{
An object of class \code{integer} of length 22.
}
\usage{
ExtensionCodes
}
\description{
NIfTI extension codes
}
\keyword{datasets}
