% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_class.R
\name{is.niftiImage}
\alias{is.niftiImage}
\alias{is.internalImage}
\title{Logical Checks on Nifti Image classes}
\usage{
is.niftiImage(x)

is.internalImage(x)
}
\arguments{
\item{x}{An object of class \code{niftiImage} or \code{internalImage}}
}
\value{
A logical \code{TRUE} or \code{FALSE}
}
\description{
Logical Checks on Nifti Image classes
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
is.niftiImage(file)
img = readNifti(file)
is.niftiImage(img)
is.internalImage(img)
int_img = readNifti(file, internal = TRUE)
is.niftiImage(int_img)
is.internalImage(int_img)
}
