% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{ndim}
\alias{ndim}
\title{Number of dimensions}
\usage{
ndim(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
The dimensionality of the object. Objects without a \code{dim}
  attribute will produce zero.
}
\description{
This function is shorthand for \code{length(dim(object))}.
}
\author{
Jon Clayden <code@clayden.org>
}

