% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti.R
\name{niftiHeader}
\alias{niftiHeader}
\alias{dumpNifti}
\title{Dump or construct a raw NIfTI header}
\usage{
niftiHeader(image = list())
}
\arguments{
\item{image}{An image, in any acceptable form (see
\code{\link{retrieveNifti}}). A list containing partial header information
is acceptable, including an empty list, which returns defaults for every
field.}

\item{x}{A \code{"niftiHeader"} object.}

\item{...}{Ignored.}
}
\value{
For \code{dumpNifti}, a list of class \code{"niftiHeader"}, with
  named components corresponding to the elements in a raw NIfTI-1 header.
}
\description{
This function extracts the contents of a NIfTI-1 header, closely
approximating how it would be stored on disk. Defaults will be used where
information is missing, but no processing is performed on the metadata.
}
\examples{
niftiHeader(system.file("extdata", "example.nii.gz", package="RNifti"))

# Default header for a standard R array
niftiHeader(array(0L, dim=c(10,10)))

}
\references{
The NIfTI-1 standard (\url{http://www.nitrc.org/docman/view.php/26/64/nifti1.h}).
}
\author{
Jon Clayden <code@clayden.org>
}
