% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nii2oro.R
\name{nii2oro}
\alias{nii2oro}
\title{Convert niftiImage to oro.nifti nifti object}
\usage{
nii2oro(image)
}
\arguments{
\item{image}{nifitImage object}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Converts a niftiImage from RNifti to a 
nifti object from the oro.nifti package
}

