% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti.R
\name{readNifti}
\alias{readNifti}
\title{Read a NIfTI-1 format file}
\usage{
readNifti(file, source = NULL, target = NULL, internal = FALSE)
}
\arguments{
\item{file}{A character vector of file names.}

\item{source, target}{If the specified \code{file} contains a transformation,
these parameters can be used to specify the associated source and target
images, which are stored in attributes of the same name. Only used if
\code{file} is of unit length.}

\item{internal}{Logical value. If \code{FALSE} (the default), an array
of class \code{"niftiImage"}, containing the image pixel or voxel values,
will be returned. If \code{TRUE}, the return value will be an object of
class \code{"internalImage"}, which contains only minimal metadata about
the image. Either way, the return value has an attribute which points to a
C data structure containing the full image.}
}
\value{
An array or internal image, with class \code{"niftiImage"}, and
  possibly also \code{"internalImage"}.
}
\description{
This function reads one or more NIfTI-1 files into R, using the standard
NIfTI-1 C library.
}
\author{
Jon Clayden <code@clayden.org>
}
\references{
The NIfTI-1 standard (\url{http://nifti.nimh.nih.gov/nifti-1}).
}
\seealso{
\code{\link{writeNifti}}
}

