% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{view}
\alias{view}
\alias{layer}
\title{A basic 3D image viewer}
\usage{
view(..., point = NULL, radiological = FALSE,
  interactive = base::interactive(), crosshairs = TRUE,
  labels = TRUE, infoPanel = defaultInfoPanel)

layer(image, scale = "grey", min = NULL, max = NULL)
}
\arguments{
\item{...}{One or more images, or \code{"viewLayer"} objects, to display.}

\item{point}{A numeric vector giving the location to initially centre the
view on. If crosshairs are in use, they will be placed at this point. For
3D images, this parameter also determines the planes shown in each
subview.}

\item{radiological}{Logical value. If \code{TRUE}, images will be displayed
in the radiological convention whereby the left of the image is shown on
the right; otherwise left is on the left.}

\item{interactive}{Logical value. If \code{TRUE}, the user can navigate
around the image by repeatedly clicking on a new centre point; otherwise
the view is fixed.}

\item{crosshairs}{Logical value, indicating whether crosshairs should be
shown or not.}

\item{labels}{Logical value, indicating whether orientation labels should be
shown or not. Ignored (defaulting to \code{FALSE}) if the image is 2D or
orientation information is not available.}

\item{infoPanel}{A function of three arguments, which must produce a plot
for the information panel of the view. \code{\link{defaultInfoPanel}} is
the default, which shows the labels and values of each image at the
current point.}

\item{image}{The image being shown in this layer.}

\item{scale}{A character vector of colour values for the scale, or a single
string naming a predefined scale: \code{"grey"} or \code{"gray"} for
greyscale, \code{"heat"} for a heatmap, \code{"rainbow"} for a rainbow
scale, or any of the scales defined in the \code{shades} package (see
\code{?shades::gradient}, if that package is installed). Ignored for RGB
images.}

\item{min, max}{The window minimum and maximum for the layer, i.e., the black
and white points. These are ignored for RGB images. Otherwise, if
\code{NULL}, the default, they are taken from the \code{cal_min} or
\code{cal_max} NIfTI header fields. If either is \code{NA}, the image has
no window stored in its header, or the two values are equal, then the 1st
and 99th percentiles of the data are used, with values close to zero
rounded to that extreme.}
}
\value{
\code{layer} returns a list of class \code{"viewLayer"}, to be used
  in a view. \code{view} is called for its side-effect of showing a view.
}
\description{
This function displays one or more 2D or 3D images, with optional
click-to-navigate interactivity.
}
\examples{
im <- readNifti(system.file("extdata", "example.nii.gz", package="RNifti"))
view(im, interactive=FALSE)

}
\seealso{
\code{\link{defaultInfoPanel}}, \code{\link{orientation}}
}
\author{
Jon Clayden <code@clayden.org>
}
